Update error messages and console logging method in server.js
Made adjustments to error messaging in server.js to provide more emphatic direction to provide a valid Mongo DB URI. Additionally, the error logging for a failed connection to Mongo DB was switched from console.log to console.error. This is a more semantically correct use of console. This will not only improve readability but also better categorize log outputs for debugging purposes.